Create shared base class for design stages
